Create a type script type definition file for fine-uploader-wrappers
Create a type script type definition file for fine-uploader-wrappers, I have this code working for me in typescript. But as it dont come with npm install, the build process in my project overwrites all the manually added files. It will be great if you can include the type definition file in npm. Feel free to extend it to include other definitions.